backend/pyproject.toml
[project]
name = "reddchat"
version = "0.1.0"
description = "Reddit AI Analysis Platform Project"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "flask>=3.0.0",
    "openai>=1.66.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]
test = [
    "pytest>=7.3.0",
    "pytest-mock>=3.10.0",
]

[build-system]
requires = ["setuptools>=65.0.0", "wheel>=0.40.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*"
addopts = "--import-mode=importlib"

[tool.setuptools]
packages = ["reddchat"]

